+++ src/components/EnterpriseApp/index.jsx
@@ -16,7 +16,7 @@ import { SCHOLAR_THEME } from '../settings/data/constants';
 import NotFoundPage from '../NotFoundPage';
 import EnterpriseAppContent from './EnterpriseAppContent';
 import { withLocation, withParams } from '../../hoc';
-import { GlobalContext } from '../GlobalProvider';
+import { GlobalContext } from '../GlobalContextProvider';
 
 class EnterpriseApp extends React.Component {
   // eslint-disable-next-line react/static-property-placement
@@ -126,7 +126,7 @@ class EnterpriseApp extends React.Component {
         enablePortalLearnerCreditManagementScreen={enablePortalLearnerCreditManagementScreen}
       >
         <BrandStyles enterpriseBranding={enterpriseBranding} />
-        <div className="enterprise-app" style={{ minHeight: this.context.minHeight }}>
+        <div className="enterprise-app fill-vertical-space" style={{ minHeight: this.context.minHeight }}>
           <MediaQuery minWidth={breakpoints.large.minWidth}>
             {matchesMediaQ => (
               <>
