+++ src/data/actions/global.js
@@ -3,20 +3,17 @@ import {
   FOOTER_HEIGHT,
 } from '../constants/global';
 
-const headerHeight = (refValue = 0) => ({
+const setHeaderHeight = (refValue = 0) => ({
   type: HEADER_HEIGHT,
   payload: { headerHeight: refValue },
 });
 
-const footerHeight = (refValue = 0) => ({
+const setFooterHeight = (refValue = 0) => ({
   type: FOOTER_HEIGHT,
   payload: { footerHeight: refValue },
 });
 
-const getGlobalHeightFromState = (state) => state;
-
 export {
-  headerHeight,
-  footerHeight,
-  getGlobalHeightFromState,
+  setHeaderHeight,
+  setFooterHeight,
 };
